Remove temp login page, redirect to dashboard from root.
